improved tests (speed and coverage)
