+++ tests/utils.py
@@ -2,6 +2,11 @@ import os
 import shutil
 import cyvcf2
 import numpy as np
+from contextlib import contextmanager
+from kipoi_utils.utils import _call_command
+
+from kipoi.cli.env import *
+from kipoi_conda.utils import *
 
 
 def compare_vcfs(fpath1, fpath2):
@@ -24,3 +29,82 @@ def cp_tmpdir(example, tmpdir):
     tdir = os.path.join(str(tmpdir), example, str(uuid4()))
     shutil.copytree(example, tdir)
     return tdir
+
+
+
+
+def create_model_env(model,source,tmpdir=None):
+    if isinstance(model, str):
+        model = [model]
+    import uuid
+
+    class Args(object):
+        def __init__(self, model, source):
+            self.model = model
+            self.source = source        
+        def _get_kwargs(self):
+            kwargs = {"dataloader": [], "env":None, "gpu": False, "model": self.model, "source": self.source,
+                  "tmpdir": "something", "vep": False}
+            return kwargs
+
+    # create the tmp dir
+    if tmpdir is None:
+        tmpdir = "/tmp/kipoi/envfiles/" + str(uuid.uuid4())[:8]
+    else:
+        tmpdir = args.tmpdir
+    if not os.path.exists(tmpdir):
+        os.makedirs(tmpdir)
+
+    # write the env file
+    logger.info("Writing environment file: {0}".format(tmpdir))
+
+
+    env, env_file = export_env(model,
+                               None,
+                               source,
+                               env_file=None,
+                               env_dir=tmpdir,
+                               env=None,
+                               vep=False,
+                               interpret=False,
+                               gpu=False)
+
+    args = Args(model=model, source=source)
+    env_db_entry = generate_env_db_entry(args, args_env_overload=env)
+    envdb = get_model_env_db()
+    envdb.append(env_db_entry)
+    envdb.save()
+
+    # setup the conda env from file
+    kipoi_conda.create_env_from_file(env_file)
+    env_db_entry.successful = True
+
+    # env is environment name
+    env_db_entry.cli_path = get_kipoi_bin(env)
+    get_model_env_db().save()
+
+
+
+
+
+def create_env_if_not_exist(model,  source, bypass=False):
+    if not bypass:
+        env_name = get_env_name(model,source=source)
+    else:
+        env_name = None
+
+    # check if we already have that env
+    if not env_exists(env_name):
+        # create env and register hook to delete env later
+        create_model_env(model=model, source=source)
+
+        import atexit
+        def call_atexit(env_name):
+            try:
+                args = ["env","remove","-n",env_name]
+                _call_command('conda' ,extra_args=args)
+            except:
+                pass
+        atexit.register(call_atexit, env_name=env_name)
+
+    return env_name
